backend/requirements.txt
fastapi
uvicorn
python-multipart
sqlalchemy
alembic
psycopg2-binary
celery
redis
anthropic
pymupdf
pdfplumber
pytesseract
pillow
python-dotenv
fhir.resources
httpx
pydantic
